requirements.txt
openai
langchain
streamlit
python_dotenv
tiktoken